Settings split with inheritence, fixed settings.
